refactor(front): representDuration() in compasable
Move representDuration() function from a SFC vue to a composable in
order to make it easier to unit test.
