+++ frontend/src/views/ReservationsView.vue
@@ -9,10 +9,10 @@
 <script setup lang="ts">
 import ClusterMainLayout from '@/components/ClusterMainLayout.vue'
 import { useClusterDataPoller } from '@/composables/DataPoller'
-import type { ClusterReservation, ClusterOptionalNumber } from '@/composables/GatewayAPI'
+import type { ClusterReservation } from '@/composables/GatewayAPI'
+import { representDuration } from '@/composables/TimeDuration'
 import InfoAlert from '@/components/InfoAlert.vue'
 import ErrorAlert from '@/components/ErrorAlert.vue'
-
 import { XMarkIcon } from '@heroicons/vue/24/outline'
 
 const props = defineProps({
@@ -23,33 +23,6 @@ const props = defineProps({
 })
 
 const { data, unable } = useClusterDataPoller<ClusterReservation[]>('reservations', 10000)
-
-function representDuration(start: ClusterOptionalNumber, end: ClusterOptionalNumber): string {
-  if (!start.set || !end.set) return '-'
-
-  let duration = end.number - start.number
-
-  let result = ''
-  if (duration > 3600 * 24) {
-    const nb_days = Math.floor(duration / (3600 * 24))
-    result += nb_days + ' day' + (nb_days > 1 ? 's' : '') + ' '
-    duration -= nb_days * 3600 * 24
-  }
-  if (duration > 3600) {
-    const nb_hours = Math.floor(duration / 3600)
-    result += nb_hours + ' hour' + (nb_hours > 1 ? 's' : '') + ' '
-    duration -= nb_hours * 3600
-  }
-  if (duration > 60) {
-    const nb_minutes = Math.floor(duration / 60)
-    result += nb_minutes + ' minute' + (nb_minutes > 1 ? 's' : '') + ' '
-    duration -= nb_minutes * 60
-  }
-  if (duration > 0) {
-    result += duration + ' second' + (duration > 1 ? 's' : '')
-  }
-  return result
-}
 </script>
 
 <template>
